checkin

+++ cpap_sounder.pl
@@ -51,7 +51,8 @@ while (1)
   {
     $date = `date`;
     chop $date;
-    play_song(10,60);
+    my $return_value = play_song(10,60);
+    next if ( $return_value eq 1 );
     print "Rebooting now at: $date size : $size to verify speaker is working\n";
     system ("touch /var/log/cpap_sounder.touch; pkill -9 play; sleep 12; shutdown -rf now");
   }
@@ -60,7 +61,8 @@ while (1)
     $date = `date`;
     chop $date;
     print "Possibly off head but still pumping air: $date size : $size\n";
-    play_song(10,60);
+    my $return_value = play_song(10,60);
+    next if ( $return_value eq 1 );
     print "Rebooting now at: $date size : $size to verify speaker is working\n";
     system ("touch /var/log/cpap_sounder.touch; pkill -9 play; sleep 12; shutdown -rf now");
   }
@@ -74,9 +76,24 @@ sub play_song
   {
     my $date = `date`;
     chop $date;
-    system ("AUDIODEV=hw:2 play alarm1.mp3 > /dev/null 2>&1");
     print "Playing song time # $count at $date\n";
-    system ("sleep $seconds_between_play");
+    system ("AUDIODEV=hw:2 play alarm1.mp3 > /dev/null 2>&1");
+    system ("sleep 10");
+    # record $seconds_between_play of audio and check if cpap is back on
+    system ("AUDIODEV=hw:1 rec small_file_size.wav trim 0 $seconds_between_play > /dev/null 2>&1 ");
+    system ("gzip --best -f small_file_size.wav");
+    my $size = `stat -c %s file_size.wav.gz`;
+    chop $size;
     $count++;
+    my $file_size_calculated = 42612996 * ($seconds_between_play / 60 / 10);
+    print "$date: Small file size: $size seconds_between_play:$seconds_between_play file_size_calculated: $file_size_calculated\n";
+
+    if ( $size < $file_size_calculated )  
+    {
+      # Will next out of while above if cpap is back on to save the reboot
+      # This makes it so we don't have to turn off/on the rasberry pi
+      return 1;
+    }
   }
+  return 0;
 }
